Move non-IO stuff back out of io/
