+++ src/robot.c
@@ -36,6 +36,7 @@
 #include "graphics.h"
 #include "idarray.h"
 #include "legacy_rasm.h"
+#include "memcasecmp.h"
 #include "robot.h"
 #include "rasm.h"
 #include "scrdisp.h"
@@ -45,7 +46,6 @@
 #include "world.h"
 #include "world_format.h"
 #include "world_struct.h"
-#include "io/memcasecmp.h"
 #include "io/memfile.h"
 #include "io/zip.h"
